Add files via upload
LAN network showcasing sub interfacing via ‘router on a stick’, VLAN configuration to help network segmentation. Redundancy across switches with LACP and routers using HSRP. Cisco's packet tracer was the network simulator used for this project, the packet tracer file for the project is included in the uploaded zipped folder.